Update DateTime and create modules for functions, object and elements
