+++ assets/css/style.css
@@ -1,6 +1,5 @@
 @import url('https://fonts.googleapis.com/css2?family=Lato:wght@100;400;700;900&display=swap');
 
-
 :root {
   --navLinksColor: #00f;
   --elementsColor: #000;
