+++ assets/js/modules/navLinksData.js
@@ -1,0 +1,2 @@
+export const navBarBtnsId = ['List', 'Add new', 'Contact'];
+export const titlesSection = ['All awesome books', 'Add a new book', 'Contact information'];
