+++ index.html
@@ -5,6 +5,7 @@
   <meta http-equiv="X-UA-Compatible" content="IE=edge">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <link rel="stylesheet" href="assets/css/style.css">
+  <script src="node_modules/luxon/build/global/luxon.js"></script>
   <title>Awesome books</title>
 </head>
 <body>
@@ -47,12 +48,12 @@
       </ul>
     </section>
     <footer class="footer-section">
-      <div class="footer-rights"> Copyright @ 2023 Powered by Ivan and Didy. All Rights Reserved. </div>  
+      <div class="footer-rights"> Copyright @ 2023 Powered by Ivan. All Rights Reserved. </div>  
     </footer>
 
   </section>
 
   </main>
-    <script src="assets/js/script.js"></script>
+    <script type="module" src="assets/js/index.js"></script>
 </body>
 </html>
